worked on PR comments. Using Key.implicit_account(), removing comments, adding contract download instructions.
